Added participants section and table, closes #29.
- Added source to generate participants table.
- Added the generated table to the repo.
- Moved functions from py files into new utils.py.
- Added small text section in paper for participants info and embedded the
  table.
