+++ paper.tex
@@ -73,9 +73,29 @@ gait data sets and may be rich enough for control identification. The data can
 also be used for verification purposes for controllers that have been designed
 in other manners.
 
-We collected over five and half hours of perturbed gait data from 11 subjects
+We collected over five and half hours of perturbed gait data from 15 subjects
 sampled at 100 hz. The data has been organized and made available for other
 research uses.
+
+\section*{Particpants}
+%
+Fifteen able bodied subjects including four females and eleven males with an
+average age of $24\pm4$ years, height of $1.75\pm0.09$ m, mass of $74\pm13$ kg
+particpated in the study. The data has been anonymized with respect to the
+participants' identities and an unique identification number was assigned to
+each subject. A selection of the meta data collected for each subject is shown
+in Table~\ref{tab:subjects}.
+%
+\begin{table*}
+  \cprotect\caption{Information about the 15 participants. The final three
+    columns give the trial numbers associated with each nominal treadmill
+    speed. Additional trials with subject 0 are trials with no subject, i.e.
+    unloaded trials that can be used for force compensation purposes, and are
+    not shown in the table. Generated by \verb|src/subject_table.py|.}
+  \centering
+  \input{tables/subjects.tex}
+  \label{tab:subjects}
+\end{table*}
 
 \section*{Equipment}
 %
