+++ src/unperturbed_perturbed_comparison.py
@@ -7,130 +7,10 @@ from numpy import rad2deg
 from scipy.constants import golden
 import matplotlib.pyplot as plt
 import pandas
-from gaitanalysis.motek import DFlowData, markers_for_2D_inverse_dynamics
-from gaitanalysis.gait import GaitData
 from gaitanalysis.utils import _percent_formatter
 import seaborn as sbn
 
-
-def trial_file_paths(trials_dir, trial_number):
-    """Returns the most common paths to the trials in the gait
-    identification data set.
-
-    Parameters
-    ==========
-    trials_dir : string
-        The path to the main directory for the data. This directory should
-        contain subdirectories: `T001/`, `T002/`, etc.
-    trial_number : string
-        Three digit trial number, e.g. `005`.
-
-    """
-
-    trial_dir = 'T' + trial_number
-    mocap_file = 'mocap-' + trial_number + '.txt'
-    record_file = 'record-' + trial_number + '.txt'
-    meta_file = 'meta-' + trial_number + '.yml'
-
-    mocap_file_path = os.path.join(trials_dir, trial_dir, mocap_file)
-    record_file_path = os.path.join(trials_dir, trial_dir, record_file)
-    meta_file_path = os.path.join(trials_dir, trial_dir, meta_file)
-
-    return mocap_file_path, record_file_path, meta_file_path
-
-
-def load_data(event, paths, tmp):
-    """Loads an event and processes the data, if necessary, from a trial
-    into a GaitData object.
-
-    Parameters
-    ==========
-    event : string
-        A valid event for the given trial.
-    paths : list of strings
-        The paths to the mocap, record, and meta data files.
-    tmp : string
-        A path to a temporary directory in which the processed data can be
-        stored.
-
-    Returns
-    =======
-    gait_data : gaitanalysis.gait.GaitData
-        The GaitData instance containing the data for the event.
-
-    """
-
-    file_name = '_'.join([n.lower() for n in event.split(' ')]) + '.h5'
-
-    tmp_data_path = os.path.join(tmp, file_name)
-
-    try:
-        f = open(tmp_data_path, 'r')
-    except IOError:
-        print('Cleaning and processing {} data...'.format(event))
-        # Load raw data, clean it up, and extract the perturbation section.
-        dflow_data = DFlowData(*paths)
-        dflow_data.clean_data(ignore_hbm=True)
-        perturbed_df = \
-            dflow_data.extract_processed_data(event=event,
-                                              index_col='TimeStamp',
-                                              isb_coordinates=True)
-
-        # Compute the lower limb 2D inverse dynamics, identify right heel
-        # strike times, and split the data into gait cycles.
-        gait_data = GaitData(perturbed_df)
-        marker_set = dflow_data.meta['trial']['marker-set']
-        subject_mass = dflow_data.meta['subject']['mass']
-        labels = markers_for_2D_inverse_dynamics(marker_set)
-        args = list(labels) + [subject_mass, 6.0]
-        gait_data.inverse_dynamics_2d(*args)
-        gait_data.grf_landmarks('FP2.ForY', 'FP1.ForY',
-                                filter_frequency=10.0,
-                                threshold=27.0)
-        gait_data.split_at('right', num_samples=80,
-                           belt_speed_column='RightBeltSpeed')
-        if not os.path.exists(tmp):
-            os.makedirs(tmp)
-        gait_data.save(tmp_data_path)
-    else:
-        print('Loading processed {} data from file...'.format(event))
-        f.close()
-        gait_data = GaitData(tmp_data_path)
-
-    return gait_data
-
-
-def remove_bad_gait_cycles(gait_data, lower, upper, col):
-    """Returns the gait cycles with outliers removed based on the
-    gait_cycle_stats DataFrame column.
-
-    Parameters
-    ==========
-    gait_data : gaitanalysis.gait.GaitData
-        The data object containing both the gait_cycles Panel and
-        gait_cycle_stats DataFrame.
-    lower : int or float
-        The lower bound for the gait_cycle_stats histogram.
-    upper : int or float
-        The upper bound for the gait_cycle_stats histogram.
-    col : string
-        The column in gait_cycle_stats to use for the bounding.
-
-    Returns
-    =======
-    gait_cycles : Panel
-        A reduced Panel of gait cycles.
-    gait_cycle_data : DataFrame
-        A reduced DataFrame of gait cycle data.
-
-    """
-
-    valid = gait_data.gait_cycle_stats[col] < upper
-    lower_values = gait_data.gait_cycle_stats[valid]
-    valid = lower_values[col] > lower
-    mid_values = lower_values[valid]
-
-    return gait_data.gait_cycles.iloc[mid_values.index], mid_values
+from utils import trial_file_paths, load_data, remove_bad_gait_cycles
 
 
 if __name__ == '__main__':
