+++ tables/subjects.tex
@@ -1,0 +1,22 @@
+\begin{tabular}{rlrrrlll}
+\toprule
+ Id &  Gender &  Age [yr] & Height [m] & Mass [kg] & 0.8 m/s &  1.2 m/s & 1.6 m/s \\
+\midrule
+    &         &           &            &           &         &          &         \\
+  1 &    male &        25 &       1.87 &       101 &      NA &  6, 7, 8 &      NA \\
+  3 &  female &        32 &       1.62 &        60 &      46 &       47 &      48 \\
+  4 &    male &        30 &       1.76 &        74 &  12, 15 &       13 &      14 \\
+  5 &    male &        23 &       1.73 &        65 &      32 &       31 &      33 \\
+  6 &    male &        26 &       1.77 &        80 &      40 &       41 &      42 \\
+  7 &  female &        29 &       1.72 &        63 &      16 &       17 &      18 \\
+  8 &    male &        20 &       1.57 &        70 &      19 &       20 &      21 \\
+  9 &    male &        20 &       1.69 &        64 &      25 &       26 &      27 \\
+ 10 &    male &        19 &       1.77 &        91 &      61 &       62 &      63 \\
+ 11 &    male &        22 &       1.85 &        80 &       9 &       10 &      11 \\
+ 12 &    male &        22 &       1.85 &        81 &      49 &       50 &      51 \\
+ 13 &  female &        21 &       1.70 &        64 &      55 &       56 &      57 \\
+ 15 &    male &        22 &       1.83 &        79 &      67 &       68 &      69 \\
+ 16 &  female &        28 &       1.69 &        52 &      76 &       77 &      78 \\
+ 17 &    male &        23 &       1.86 &        87 &      73 &       74 &      75 \\
+\bottomrule
+\end{tabular}
